[Add] 섬 시스템 기본 틀 제작 + 보드게임씬 섬 정적할당
1) 섬 시스템 기본 틀 제작
2) 보드게임씬에서 각 섬에서 이어지는 다음과 이전 섬 정적할당 완료
