+++ Team_Immortal Sprouts_Pummel Party/Assets/Scenes/WooSeok/BoardGameWooseok.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a114f51f6c442e34a9415b6f8ed6fd4d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
