--- Team_Immortal Sprouts_Pummel Party/Assets/Scripts/BaseShark.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: febb0a43ab18a624f904d70fdea6654e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
