+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Fish/FishRotate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 142b741404c9e464187e6e554c1ca4c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
