+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/BasicIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6250551db057dcd4c83d6cbf428abed0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
