+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/HealIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 825f9d0a0861fb84c8fc666ab42f0c4b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
