+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/ItemIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ce3f3506313f424897fc42b81e9fe88
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
