+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/NobodyIsland.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NobodyIsland : Island
+{
+    [SerializeField] Transform[] nobodyIslandPlayerSpots;
+    
+    void Start()
+    {
+        InitPositionSettings().Forget();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
