+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/NobodyIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b78bce222a2d194ab7a0abe58e21c13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
