+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/RotationIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6e43f238b3a4284baf7e98a0d10992b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
