+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/SharkIsland.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SharkIsland : Island
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        InitPositionSettings().Forget();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
