+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/TrophyIsland.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TrophyIsland : Island
+{
+    void Start()
+    {
+        InitPositionSettings().Forget();
+    }
+
+    void Update()
+    {
+        
+    }
+}
