+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Islands/TrophyIsland.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db291f9a46ecbd844aba1f17a72531dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
