--- Team_Immortal Sprouts_Pummel Party/Assets/Scripts/JumpShark.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e06d3e4ad4f15748990e8b3bbaffaec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
