+++ Team_Immortal Sprouts_Pummel Party/Assets/Static GameObjects/Islands/Heal Island/Heal Island.prefab
@@ -9,6 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 545757268109360300}
+  - component: {fileID: 6221676610569007539}
   m_Layer: 0
   m_Name: Heal Island
   m_TagString: Untagged
@@ -50,6 +51,21 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &6221676610569007539
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 545757268109360303}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 825f9d0a0861fb84c8fc666ab42f0c4b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  nextIsland: {fileID: 0}
+  PrevIsland: {fileID: 0}
+  currentPlayerSpot: {fileID: 545757269096143180}
 --- !u!1 &545757268710958475
 GameObject:
   m_ObjectHideFlags: 0
